Refactored the code using javierjulio's code as inspiration
JavierJulio's code is here:
https://github.com/javierjulio/ruby-koans-completed/blob/master/about_scoring_project.rb#L44

javierjulio stored the frequency of each die roll into an Hash and then looked up the frequency of the die rolls when calculating the score. This was much better than the old method of checking the array by using a "for-loop". javierjulio's way reduced the length of my code by a few lines. I added one minor improvement though:
*Deleting ALL scored rolls from the Hash, potentially allowing for an expansion of this code to include the feature of rerolling. In the standard version of "Greed", you are allowed to reroll any die that were not scored. Erasing all dice that were scored (thereby keeping track of unscored dice) could be an important step in including the feature of rerolling. (No, I will not actually implement rerolling. There's just so few dice remaining that rerolling would not be worthwhile. The standard version of "Greed" had you rolling 10 dice, not 5, so rerolling made sense there.)